pyproject.toml
[project]
name = "llm-orchestration-benchmark"
version = "0.1.0"
description = "A comprehensive benchmark for evaluating LLM orchestration capabilities"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    {name = "Sooyeon Lee", email = "soo.x@kakaocorp.com"}
]
license = {text = "MIT"}
keywords = ["llm", "benchmark", "orchestration", "multi-agent", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "openai>=1.95.1",
    "anthropic>=0.57.1",
    "google-generativeai>=0.8.5",
    "google-genai>=1.27.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "uvicorn[standard]>=0.35.0",
    "loguru>=0.7.3",
    "typer>=0.16.0",
    "rich>=14.0.0",
    "pandas>=2.3.1",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "plotly>=6.2.0",
    "jinja2>=3.1.6",
    "tenacity>=9.1.2",
    "asyncio-throttle>=1.0.2",
    "gspread>=6.2.1",
    "networkx>=3.5",
    "httpx>=0.28.1",
    "boto3>=1.40.2",
    "botocore>=1.40.2",
    "scipy>=1.16.1",
    "ruamel.yaml>=0.18.0",
    "huggingface-hub>=0.25.0",
    "typer>=0.21.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
vllm = [
    "vllm>=0.11.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "poethepoet>=0.33.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.2",
    "types-pyyaml>=6.0.12.20250402",
    "types-requests>=2.31.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
evaluate = "integration.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "integration"]

[tool.poe.tasks]
# Formatting
_black-check = "black --check --line-length 100 src/ tests/ scripts/"
_isort-check = "isort --check-only src/ tests/ scripts/"
formatting-check = ["_black-check", "_isort-check"]
_black = "black --line-length 100 src/ tests/ scripts/"
_isort = "isort src/ tests/ scripts/"
format = ["_black", "_isort"]

# Linting
lint = "ruff check src/ tests/ scripts/"
typetest = "mypy src/ tests/ scripts/"

# Testing
test = "pytest tests/ -v"
test-cov = "pytest tests/ --cov=src --cov-report=html"

# Development
[tool.poe.tasks.serve]
cmd = "python run.py serve"
help = "Start the multi-agent orchestration server"

[tool.poe.tasks.serve-debug]
cmd = "python run.py serve --debug"
help = "Start the server with debug logging"

[tool.poe.tasks.serve-custom]
cmd = "python run.py serve --config ${config} --agent-cards ${cards} --port ${port}"
help = "Start server with custom configuration"

[tool.poe.tasks.list-agents]
cmd = "python run.py list-agents"
help = "List all available agents"

[tool.poe.tasks.validate-config]
cmd = "python run.py validate-config"
help = "Validate configuration files"

[tool.poe.tasks.test-llm]
cmd = "python run.py test-llm"
help = "Test LLM configuration"

# Legacy tasks (keeping for backward compatibility)
[tool.poe.tasks.devserver]
cmd = "python run.py serve --debug"
help = "Legacy: Start development server"

[tool.poe.tasks.benchmark]
cmd = "python run.py serve"
help = "Legacy: Start server"

# Setup
[tool.poe.tasks.setup-dev]
sequence = [
    {shell = "pre-commit install"},
    {ref = "format"},
    {ref = "test"}
]

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "C",   # mccabe
    "B",   # flake8-bugbear
    "T20", # flake8-print
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--color=yes",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
